LPS-206027 Move logic of Permissions to a new component
